# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TERA EXAM Sarkari exam platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthToken"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current admin profile
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Current admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admin"
        "401":
          description: Not authenticated

  # ---------------------------------------------------------------------
  # Public: home aggregate
  # ---------------------------------------------------------------------
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List published jobs
      parameters:
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Jobs
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Job" }
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a published job
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Job
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
        "404":
          description: Not found

  # ---------------------------------------------------------------------
  # Public: Results
  # ---------------------------------------------------------------------
...
  /admit-cards:
    get:
      operationId: listAdmitCards
      tags: [admit-cards]
      summary: List published admit cards
      parameters:
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Admit cards
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AdmitCard" }
  /admit-cards/{id}:
    get:
      operationId: getAdmitCard
      tags: [admit-cards]
      summary: Get a published admit card
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Admit card
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdmitCard" }
        "404":
          description: Not found

  # ---------------------------------------------------------------------
  # Public: Syllabus
  # ---------------------------------------------------------------------
...
  /admin/jobs:
    get:
      operationId: adminListJobs
      tags: [admin]
      summary: List all jobs (any status)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/PublishStatus" }
      responses:
        "200":
          description: Jobs
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Job" }
    post:
      operationId: createJob
      tags: [admin]
      summary: Create a job notification
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JobInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
  /admin/jobs/{id}:
    patch:
      operationId: updateJob
      tags: [admin]
      summary: Update a job (also used for one-click publish via status field)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JobUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
        "404":
          description: Not found
    delete:
      operationId: deleteJob
      tags: [admin]
      summary: Delete a job
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ---------------------------------------------------------------------
  # Admin: Results
  # ---------------------------------------------------------------------
...
  /admin/admit-cards:
    get:
      operationId: adminListAdmitCards
      tags: [admin]
      summary: List all admit cards (any status)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/PublishStatus" }
      responses:
        "200":
          description: Admit cards
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AdmitCard" }
    post:
      operationId: createAdmitCard
      tags: [admin]
      summary: Create an admit card link
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdmitCardInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdmitCard" }
  /admin/admit-cards/{id}:
    patch:
      operationId: updateAdmitCard
      tags: [admin]
      summary: Update an admit card (also used for one-click publish via status field)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdmitCardUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdmitCard" }
        "404":
          description: Not found
    delete:
      operationId: deleteAdmitCard
      tags: [admin]
      summary: Delete an admit card
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ---------------------------------------------------------------------
  # Admin: Syllabus
  # ---------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PublishStatus:
      type: string
      enum: [draft, published]
...
    Admin:
      type: object
      required: [id, username, createdAt]
      properties:
        id: { type: integer }
        username: { type: string }
        createdAt: { type: string, format: date-time }
    AdminLoginInput:
      type: object
      required: [username, password]
      properties:
        username: { type: string, minLength: 1 }
        password: { type: string, minLength: 1 }
    AuthToken:
      type: object
      required: [token, admin]
      properties:
        token: { type: string }
        admin: { $ref: "#/components/schemas/Admin" }

    # --- Job ----------------------------------------------------------
    Job:
      type: object
      required:
        [id, title, organization, examName, postDate, status, slug, createdAt, updatedAt]
      properties:
        id: { type: integer }
        title: { type: string }
        organization: { type: string }
        examName: { type: string }
        postDate: { type: string, format: date }
        lastDate: { type: ["string", "null"], format: date }
        applyLink: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        vacancies: { type: ["integer", "null"] }
        category: { type: ["string", "null"] }
        status: { $ref: "#/components/schemas/PublishStatus" }
        slug: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    JobInput:
      type: object
      required: [title, organization, examName, postDate, slug]
      properties:
        title: { type: string, minLength: 1 }
        organization: { type: string, minLength: 1 }
        examName: { type: string, minLength: 1 }
        postDate: { type: string, format: date }
        lastDate: { type: ["string", "null"], format: date }
        applyLink: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        vacancies: { type: ["integer", "null"] }
        category: { type: ["string", "null"] }
        status: { $ref: "#/components/schemas/PublishStatus" }
        slug: { type: string, minLength: 1 }
    JobUpdate:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        organization: { type: string, minLength: 1 }
        examName: { type: string, minLength: 1 }
        postDate: { type: string, format: date }
        lastDate: { type: ["string", "null"], format: date }
        applyLink: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        vacancies: { type: ["integer", "null"] }
        category: { type: ["string", "null"] }
        status: { $ref: "#/components/schemas/PublishStatus" }
        slug: { type: string, minLength: 1 }

    # --- Result ---------------------------------------------------------
...
    AdmitCard:
      type: object
      required:
        [id, title, organization, examName, examDate, status, slug, createdAt, updatedAt]
      properties:
        id: { type: integer }
        title: { type: string }
        organization: { type: string }
        examName: { type: string }
        examDate: { type: string, format: date }
        admitCardLink: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        status: { $ref: "#/components/schemas/PublishStatus" }
        slug: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AdmitCardInput:
      type: object
      required: [title, organization, examName, examDate, slug]
      properties:
        title: { type: string, minLength: 1 }
        organization: { type: string, minLength: 1 }
        examName: { type: string, minLength: 1 }
        examDate: { type: string, format: date }
        admitCardLink: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        status: { $ref: "#/components/schemas/PublishStatus" }
        slug: { type: string, minLength: 1 }
    AdmitCardUpdate:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        organization: { type: string, minLength: 1 }
        examName: { type: string, minLength: 1 }
        examDate: { type: string, format: date }
        admitCardLink: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        status: { $ref: "#/components/schemas/PublishStatus" }
        slug: { type: string, minLength: 1 }

    # --- Syllabus -----------------------------------------------------
...