# lib/api-zod/src/generated/types/admin.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * TERA EXAM Sarkari exam platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Admin {
  id: number;
  username: string;
  createdAt: Date;
}